Convert Wave into a sine wave calculator
Sampling is done with a new class: `WaveSampler`.

+++ src/main.rs
@@ -6,7 +6,7 @@ fn main() {
     let mut stream = WavStream::new();
 
     for frequency in &[500, 600, 400, 500, 800, 900, 1000, 800, 600, 400, 300] {
-        stream.wave(0.5, *frequency, 0.22);
+        stream.wave(0.5, *frequency as f32, 0.22);
     }
 
     stream.finish();
